Use Admin Controlled for Client and Prover
