+++ contracts/near/admin-controlled/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "admin-controlled"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+near-sdk = "0.11.0"
