+++ contracts/near/eth-client/Cargo.toml
@@ -9,6 +9,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 eth-types =  { path = "../eth-types" }
+admin-controlled =  { path = "../admin-controlled" }
 near-sdk = "0.11.0"
 borsh = "0.6.2"
 wee_alloc = { version = "0.4.5", default-features = false, features = [] }
