+++ contracts/near/eth-prover/Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 wee_alloc = { version = "0.4.5", default-features = false, features = [] }
 eth-types =  { path = "../eth-types" }
+admin-controlled =  { path = "../admin-controlled" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 near-sdk = "0.11.0"
